# Aggressive splom perf