--- a/src/plot_api/plot_api.js
+++ b/src/plot_api/plot_api.js
@@ -1701,15 +1701,26 @@ exports.relayout = function relayout(gd, astr, val) {
         seq.push(subroutines.layoutReplot);
     }
     else if(Object.keys(aobj).length) {
-        Plots.supplyDefaults(gd);
+        var flagList = Object.keys(flags).filter(function(k) { return flags[k]; });
+
+        // Optimization mostly for large splom traces where
+        // Plots.supplyDefaults can take > 100ms
+        if(flagList[0] === 'axrange' && flagList.length === 1) {
+            for(var k in specs.rangesAltered) {
+                var axName = Axes.id2name(k);
+                var axIn = gd.layout[axName];
+                var axOut = gd._fullLayout[axName];
+                axOut.autorange = axIn.autorange;
+                axOut.range = axIn.range.slice();
+                axOut.cleanRange();
+            }
+        } else {
+            Plots.supplyDefaults(gd);
+        }
 
         if(flags.legend) seq.push(subroutines.doLegend);
         if(flags.layoutstyle) seq.push(subroutines.layoutStyles);
-
-        if(flags.axrange) {
-            addAxRangeSequence(seq, specs.rangesAltered);
-        }
-
+        if(flags.axrange) addAxRangeSequence(seq, specs.rangesAltered);
         if(flags.ticks) seq.push(subroutines.doTicksRelayout);
         if(flags.modebar) seq.push(subroutines.doModeBar);
         if(flags.camera) seq.push(subroutines.doCamera);