… src/plots/cartesian/dragbox.js
… src/plots/cartesian/index.js
… src/plots/cartesian/select.js
… src/plots/plots.js
… src/plots/polar/layout_attributes.js
… src/plots/polar/polar.js
… src/plots/sort_modules.js
… src/traces/scattergl/attributes.js
… src/traces/scattergl/index.js
… src/traces/splom/attributes.js
… src/traces/splom/base_plot.js
… src/traces/splom/defaults.js
… src/traces/splom/index.js
… test/image/baselines/splom_nodiag.png
… test/image/mocks/splom_nodiag.json
… test/jasmine/assets/custom_assertions.js
… test/jasmine/tests/cartesian_test.js
… test/jasmine/tests/gl2d_click_test.js
… test/jasmine/tests/gl2d_plot_interact_test.js
… test/jasmine/tests/plot_api_test.js
… test/jasmine/tests/plots_test.js
… test/jasmine/tests/select_test.js
… test/jasmine/tests/splom_test.js
… test/jasmine/tests/streamtube_test.js