Completes the splom part of https://github.com/plotly/plotly.js/issues/2553 and checks off two O(n^2) items in https://github.com/plotly/plotly.js/issues/2555.

Each commit here is fairly self-contained. The most _aggresive_ optimizations are in https://github.com/plotly/plotly.js/commit/17e30d2f5f5bdb6296f90fdceea514b8c5dd8518 (aka `supplyDefaults`-bypass) and https://github.com/plotly/plotly.js/commit/4792f08771750a421e3e33895f602580ffa7415b (aka no-need-for-`<rect.bg>`-nor-subplot-`<clipPath>`) Note that commit https://github.com/plotly/plotly.js/commit/3574ece2a1eff74e1009055d9bb1ca699bafc1f9 was required to not have to re-init the splom objects on `editType: 'style'` updates.

In terms of performance gains, using:

```js
var Nvars = 50
var Nrows = 800
var dims = []

for(var i = 0; i < Nvars; i++) {
  dims.push({values: []})
  for(var j = 0; j < Nrows; j++) {
     dims[i].values.push(Math.random())
  }
}

var gd = document.getElementById('graph')
Plotly.purge(gd);

console.time('splom')
Plotly.newPlot(gd, [{
  type: 'splom',
  dimensions: dims,
  showupperhalf: false,
  diagonal: {visible: false}
  // totaling 1225 subplots
}], {width: 2000, height: 2000},
   {showAxisDragHandles: false, showAxisRangeEntryBoxes: false}
)
console.timeEnd('splom')
```

and comparing against `v1.41.3` on Chrome 69 (which significantly slower than Chrome 66, by the way), we get gains of:

- ~200 ms on first render, thanks to the no-need-for-`<rect.bg>`-nor-subplot-`<clipPath>`
- ~200 ms on `relayout(gd, 'xaxis.range', [/* */])`, thanks to https://github.com/plotly/plotly.js/commit/cc5f0caafe230c21b8f3f25d3b7b9b244be5ebf8 and the `supplyDefaults`-bypass
- huge improvements for `restyle(gd, 'marker.color', /* scalar */)`: ~150ms vs 1000ms
- huge improvements (but not good enough) for `restyle(gd, 'marker.size', /* scalar */)`: ~400ms vs 1000ms
- huge improvements (but not good enough) for `restyle(gd, 'marker.color', '/* arrayOk */)`: ~800ms vs 2000ms

cc @alexcjohnson @antoinerg @archmoj