@etpinard 2018-09-28
Here are some benchmarks in the splom pipeline off this branch. At 50 dims and 800 rows, we clock:

- `supplyDefaults`: 100-150 ms
- `doCalcdata`: ~20 ms
- `drawFramework` (where the `<g.subplot>` are added to the DOM): 200-300 ms
   + room for improvement here!
- `lsInner` (where subplot, bg, lines and clip paths are updated): 40-60 ms
- `drawAxes`: 300-400 ms
- `initInteractions` (where the drag layers are added to the DOM): 
   + ~100-150 ms (with `showAxisDragHandles` and `showAxisRangeEntryBoxes` false)
   + ~1000 ms (default),
- `createREGL`: ~300-400 ms on a 2000x2000px canvas
- splom grid update/draw: ~150-200 ms, 
   + we could add `dimensions[i].axis.showgrid: false` to skip this
- splom init (the `createMatrix` call): ~200 ms (not sure why)
- splom update/draw: ~400-600 ms

which gives a total range of 1800 to 2500 ms.