@alexcjohnson 2018-09-29
> - `initInteractions` (where the drag layers are added to the DOM): 
>    + ~100-150 ms (with `showAxisDragHandles` and `showAxisRangeEntryBoxes` false)
>    + ~1000 ms (default),

Wow - time to experiment with a single drag element that registers many separate regions?