@alexcjohnson 2018-09-29
Fantastic work @etpinard - very impressive results!

Just two questions of safety https://github.com/plotly/plotly.js/pull/3057#discussion_r221407615 https://github.com/plotly/plotly.js/pull/3057#discussion_r221413342 and a couple of potential small improvements https://github.com/plotly/plotly.js/pull/3057#discussion_r221406493 https://github.com/plotly/plotly.js/pull/3057#discussion_r221411382

The single drag element is already mentioned in #2555 and obviously should not be done here, that should be its own PR.