@etpinard 2018-10-03
Improvemens (e.g. no axis type clearance) over https://github.com/plotly/plotly.js/pull/3057/commits/078c0864014fb93c8ca8b0525353144a9d0eece0 in -> https://github.com/plotly/plotly.js/compare/splom-perf...splom-axis-dim-match